moved refactored code to test_single.py and added xfail as key in examples so that debugging can be done easily
