+++ sympy/solvers/ode/tests/test_ode.py
@@ -7,7 +7,7 @@ from sympy.solvers.ode import (classify_ode,
     dsolve)
 
 from sympy.solvers.ode.subscheck import checkodesol, checksysodesol
-from sympy.solvers.ode.ode import (allhints, _linear_coeff_match,
+from sympy.solvers.ode.ode import (_linear_coeff_match,
     _undetermined_coefficients_match, classify_sysode,
     constant_renumber, constantsimp, get_numbered_constants, solve_ics)
 
@@ -16,7 +16,6 @@ from sympy.functions import airyai, airybi, besselj, bessely
 from sympy.solvers.deutils import ode_order
 from sympy.testing.pytest import XFAIL, skip, raises, slow, ON_TRAVIS, SKIP
 from sympy.utilities.misc import filldedent
-import traceback
 
 C0, C1, C2, C3, C4, C5, C6, C7, C8, C9, C10 = symbols('C0:11')
 u, x, y, z = symbols('u,x:z', real=True)
@@ -32,244 +31,6 @@ h = Function('h')
 # Tests of order higher than 1 should run the solutions through
 # constant_renumber because it will normalize it (constant_renumber causes
 # dsolve() to return different results on different machines)
-
-
-hint_message = """\
-Hint did not match the example {example}.
-
-The ODE is:
-{eq}.
-
-The expected hint was
-{our_hint}\
-"""
-
-expected_sol_message = """\
-Different solution found from dsolve for example {example}.
-
-The ODE is:
-{eq}
-
-The expected solution was
-{sol}
-
-What dsolve returned is:
-{dsolve_sol}\
-"""
-
-checkodesol_msg = """\
-solution found is not correct for example {example}.
-
-The ODE is:
-{eq}\
-"""
-
-dsol_incorrect_msg = """\
-solution returned by dsolve is incorrect when using {hint}.
-
-The ODE is:
-{eq}
-
-The expected solution was
-{sol}
-
-what dsolve returned is:
-{dsolve_sol}\
-"""
-
-exception_msg = """\
-dsolve raised exception : {e}
-
-when using {hint} for the example {example}
-
-You can test this with:
-
-from sympy.solvers.ode.tests.test_ode import _test_particular_example
-
-_test_particular_example('{hint}', '{example}')
-
-The ODE is:
-{eq}\
-"""
-
-def _test_for_particular_hint(our_hint, ode_example):
-    eq = ode_example['eq']
-    expected_sol = ode_example['sol']
-    example = ode_example['example_name']
-    if our_hint in classify_ode(eq):
-        try:
-            dsolve_sol = dsolve(eq, hint=our_hint)
-            expected_checkodesol = [(True, 0) for i in range(len(expected_sol))]
-            if len(expected_sol) == 1:
-                expected_checkodesol = (True, 0)
-
-            if checkodesol(eq, dsolve_sol) != expected_checkodesol:
-                message = dsol_incorrect_msg.format(hint=our_hint, eq=eq, sol=expected_sol,dsolve_sol=dsolve_sol)
-                raise AssertionError(message)
-        except Exception as e:
-            print(exception_msg.format(e=str(e), hint=our_hint, example=example, eq=eq))
-            traceback.print_exc()
-
-
-def _ode_solver_test(ode_examples):
-    our_hint = ode_examples['hint']
-    for example in ode_examples['examples']:
-        eq = ode_examples['examples'][example]['eq']
-        sol = ode_examples['examples'][example]['sol']
-        if our_hint not in classify_ode(eq):
-            message = hint_message.format(example=example, eq=eq, our_hint=our_hint)
-            raise AssertionError(message)
-
-        dsolve_sol = dsolve(eq,hint=our_hint)
-        if dsolve_sol not in sol:
-            message = expected_sol_message.format(example=example, eq=eq, sol=sol, dsolve_sol=dsolve_sol)
-            raise AssertionError(message)
-
-        expected_checkodesol = [(True, 0) for i in range(len(sol))]
-        if checkodesol(eq, sol) != expected_checkodesol:
-            message = checkodesol.format(example=example, eq=eq)
-            raise AssertionError(message)
-
-
-def _get_examples_ode_sol_euler_homogeneous():
-    return {
-            'hint': "nth_linear_euler_eq_homogeneous",
-            'func': f(x),
-            'examples':{
-    'euler_hom_01': {
-        'eq': Eq(-3*diff(f(x), x)*x + 2*x**2*diff(f(x), x, x), 0),
-        'sol': [Eq(f(x), C1 + C2*x**Rational(5, 2))],
-    },
-
-    'euler_hom_02': {
-        'eq': Eq(3*f(x) - 5*diff(f(x), x)*x + 2*x**2*diff(f(x), x, x), 0),
-        'sol': [Eq(f(x), C1*sqrt(x) + C2*x**3)]
-    },
-
-    'euler_hom_03': {
-        'eq': Eq(4*f(x) + 5*diff(f(x), x)*x + x**2*diff(f(x), x, x), 0),
-        'sol': [Eq(f(x), (C1 + C2*log(x))/x**2)]
-    },
-
-    'euler_hom_04': {
-        'eq': Eq(6*f(x) - 6*diff(f(x), x)*x + 1*x**2*diff(f(x), x, x) + x**3*diff(f(x), x, x, x), 0),
-        'sol': [Eq(f(x), C1/x**2 + C2*x + C3*x**3)]
-    },
-
-    'euler_hom_05': {
-        'eq': Eq(-125*f(x) + 61*diff(f(x), x)*x - 12*x**2*diff(f(x), x, x) + x**3*diff(f(x), x, x, x), 0),
-        'sol': [Eq(f(x), x**5*(C1 + C2*log(x) + C3*log(x)**2))]
-    },
-
-    'euler_hom_06': {
-        'eq': x**2*diff(f(x), x, 2) + x*diff(f(x), x) - 9*f(x),
-        'sol': [Eq(f(x), C1*x**-3 + C2*x**3)]
-    },
-
-    'euler_hom_07': {
-        'eq': sin(x)*x**2*f(x).diff(x, 2) + sin(x)*x*f(x).diff(x) + sin(x)*f(x),
-        'sol': [Eq(f(x), C1*sin(log(x)) + C2*cos(log(x)))]
-    },
-    }
-    }
-
-
-def _get_examples_ode_sol_euler_undetermined_coeff():
-    return {
-            'hint': "nth_linear_euler_eq_nonhomogeneous_undetermined_coefficients",
-            'func': f(x),
-            'examples':{
-    'euler_undet_01': {
-        'eq': Eq(x**2*diff(f(x), x, x) + x*diff(f(x), x), 1),
-        'sol': [Eq(f(x), C1 + C2*log(x) + log(x)**2/2)]
-    },
-
-    'euler_undet_02': {
-        'eq': Eq(x**2*diff(f(x), x, x) - 2*x*diff(f(x), x) + 2*f(x), x**3),
-        'sol': [Eq(f(x), x*(C1 + C2*x + Rational(1, 2)*x**2))]
-    },
-
-    'euler_undet_03': {
-        'eq': Eq(x**2*diff(f(x), x, x) - x*diff(f(x), x) - 3*f(x), log(x)/x),
-        'sol': [Eq(f(x), (C1 + C2*x**4 - log(x)**2/8 - log(x)/16)/x)]
-    },
-
-    'euler_undet_04': {
-        'eq': Eq(x**2*diff(f(x), x, x) + 3*x*diff(f(x), x) - 8*f(x), log(x)**3 - log(x)),
-        'sol': [Eq(f(x), C1/x**4 + C2*x**2 - Rational(1,8)*log(x)**3 - Rational(3,32)*log(x)**2 - Rational(1,64)*log(x) - Rational(7, 256))]
-    },
-
-    'euler_undet_05': {
-        'eq': Eq(x**3*diff(f(x), x, x, x) - 3*x**2*diff(f(x), x, x) + 6*x*diff(f(x), x) - 6*f(x), log(x)),
-        'sol': [Eq(f(x), C1*x + C2*x**2 + C3*x**3 - Rational(1, 6)*log(x) - Rational(11, 36))]
-    },
-    }
-    }
-
-
-def _get_examples_ode_sol_euler_var_para():
-    return {
-            'hint': "nth_linear_euler_eq_nonhomogeneous_variation_of_parameters",
-            'func': f(x),
-            'examples':{
-    'euler_var_01': {
-        'eq': Eq(x**2*Derivative(f(x), x, x) - 2*x*Derivative(f(x), x) + 2*f(x), x**4),
-        'sol': [Eq(f(x), x*(C1 + C2*x + x**3/6))]
-    },
-
-    'euler_var_02': {
-        'eq': Eq(3*x**2*diff(f(x), x, x) + 6*x*diff(f(x), x) - 6*f(x), x**3*exp(x)),
-        'sol': [Eq(f(x), C1/x**2 + C2*x + x*exp(x)/3 - 4*exp(x)/3 + 8*exp(x)/(3*x) - 8*exp(x)/(3*x**2))]
-    },
-
-    'euler_var_03': {
-        'eq': Eq(x**2*Derivative(f(x), x, x) - 2*x*Derivative(f(x), x) + 2*f(x), x**4*exp(x)),
-        'sol':  [Eq(f(x), x*(C1 + C2*x + x*exp(x) - 2*exp(x)))]
-    },
-
-    'euler_var_04': {
-        'eq': x**2*Derivative(f(x), x, x) - 2*x*Derivative(f(x), x) + 2*f(x) - log(x),
-        'sol': [Eq(f(x), C1*x + C2*x**2 + log(x)/2 + Rational(3, 4))]
-    },
-
-    'euler_var_05': {
-        'eq': -exp(x) + (x*Derivative(f(x), (x, 2)) + Derivative(f(x), x))/x,
-        'sol': [Eq(f(x), C1 + C2*log(x) + exp(x) - Ei(x))]
-    },
-    }
-    }
-
-
-def _get_all_examples():
-    all_solvers = [_get_examples_ode_sol_euler_homogeneous(), _get_examples_ode_sol_euler_undetermined_coeff(), _get_examples_ode_sol_euler_var_para()]
-    all_examples = []
-    for solver in all_solvers:
-        for example in solver['examples']:
-            temp = {
-                'hint': solver['hint'],
-                'func': solver['examples'][example].get('func',solver['func']),
-                'eq': solver['examples'][example]['eq'],
-                'sol': solver['examples'][example]['sol'],
-                'example_name': example,
-            }
-            all_examples.append(temp)
-    return all_examples
-
-
-def _test_all_hints():
-    all_hints = list(allhints)
-    all_examples = _get_all_examples()
-    for our_hint in all_hints:
-        for ode_example in all_examples:
-            _test_for_particular_hint(our_hint, ode_example)
-
-
-def _test_particular_example(our_hint, example_name):
-    all_examples = _get_all_examples()
-    for example in all_examples:
-        if example['example_name'] == example_name:
-            eq = example['eq']
-            dsolve(eq, hint=our_hint)
 
 
 def test_get_numbered_constants():
@@ -3027,49 +2788,6 @@ def test_issue_5770():
 def test_issue_5112_5430():
     assert homogeneous_order(-log(x) + acosh(x), x) is None
     assert homogeneous_order(y - log(x), x, y) is None
-
-
-def test_nth_order_linear_euler_eq_homogeneous():
-    x, t, a, b, c = symbols('x t a b c')
-    y = Function('y')
-    our_hint = "nth_linear_euler_eq_homogeneous"
-
-    eq = diff(f(t), t, 4)*t**4 - 13*diff(f(t), t, 2)*t**2 + 36*f(t)
-    assert our_hint in classify_ode(eq)
-
-    eq = a*y(t) + b*t*diff(y(t), t) + c*t**2*diff(y(t), t, 2)
-    assert our_hint in classify_ode(eq)
-
-    _ode_solver_test(_get_examples_ode_sol_euler_homogeneous())
-
-
-
-def test_nth_order_linear_euler_eq_nonhomogeneous_undetermined_coefficients():
-    x, t = symbols('x t')
-    a, b, c, d = symbols('a b c d', integer=True)
-    our_hint = "nth_linear_euler_eq_nonhomogeneous_undetermined_coefficients"
-
-    eq = x**4*diff(f(x), x, 4) - 13*x**2*diff(f(x), x, 2) + 36*f(x) + x
-    assert our_hint in classify_ode(eq, f(x))
-
-    eq = a*x**2*diff(f(x), x, 2) + b*x*diff(f(x), x) + c*f(x) + d*log(x)
-    assert our_hint in classify_ode(eq, f(x))
-
-    _ode_solver_test(_get_examples_ode_sol_euler_undetermined_coeff())
-
-
-def test_nth_order_linear_euler_eq_nonhomogeneous_variation_of_parameters():
-    x, t = symbols('x, t')
-    a, b, c, d = symbols('a, b, c, d', integer=True)
-    our_hint = "nth_linear_euler_eq_nonhomogeneous_variation_of_parameters"
-
-    eq = Eq(x**2*diff(f(x),x,2) - 8*x*diff(f(x),x) + 12*f(x), x**2)
-    assert our_hint in classify_ode(eq, f(x))
-
-    eq = Eq(a*x**3*diff(f(x),x,3) + b*x**2*diff(f(x),x,2) + c*x*diff(f(x),x) + d*f(x), x*log(x))
-    assert our_hint in classify_ode(eq, f(x))
-
-    _ode_solver_test(_get_examples_ode_sol_euler_var_para())
 
 
 def test_issue_5095():
